requirements.txt
ultralytics
load_dotenv
numpy==1.26.4
